chore: refactor nx target graph to better include ast-spec copy
