+++ packages/types/src/index.ts
@@ -1,5 +1,4 @@
-export { AST_NODE_TYPES } from './ast-spec';
-export { AST_TOKEN_TYPES } from './ast-spec';
+export { AST_NODE_TYPES, AST_TOKEN_TYPES } from './generated/ast-spec';
 export * from './lib';
 export * from './parser-options';
 export * from './ts-estree';
