+++ packages/typescript-estree/src/convert.ts
@@ -35,7 +35,6 @@ import {
   TSNode,
 } from './ts-estree';
 import { typescriptVersionIsAtLeast } from './version-check';
-import { Expression } from '@typescript-eslint/types/src/ast-spec';
 
 const SyntaxKind = ts.SyntaxKind;
 
@@ -1925,7 +1924,7 @@ export class Converter {
             expressions: [],
           });
 
-          const left = this.convertChild(node.left) as Expression;
+          const left = this.convertChild(node.left) as TSESTree.Expression;
           if (
             left.type === AST_NODE_TYPES.SequenceExpression &&
             node.left.kind !== SyntaxKind.ParenthesizedExpression
